Split code into separate files for easier navigation.
